requirements.txt
keyboard==0.13.5
pystray==0.19.5
Pillow==10.4.0
